Port recipes, recipe displays and remaining block APIs to 1.21.8
The Recipe interface gained placementInfo, recipeBookCategory and display,
and lost getIngredients, getGroup, getResultItem, canCraftInDimensions and
getToastSymbol. Cooking pot and cutting board recipes implement the new
contract, with dedicated RecipeDisplay records carrying what clients need
(recipes themselves are no longer synchronized) and registered display types.

Recipe book categories became registry entries, so the mod registers its own
for cooking meals, drinks, misc and cutting. SimpleCraftingRecipeSerializer
was folded into CustomRecipe.Serializer.

Blocks: BushBlock gained a concrete codec and default bonemeal behaviour in
1.21.5, so blocks that relied on the old BushBlock extend VegetationBlock
instead; plus fixes for Orientation, TriState, PiglinAi.angerNearbyPiglins,
Direction.getUnitVec3i, registry holder lookups and server-only recipe access.

+++ src/main/java/vectorwing/farmersdelight/common/block/RiceBaleBlock.java
@@ -26,8 +26,8 @@ public class RiceBaleBlock extends Block
 	}
 
 	@Override
-	public void fallOn(Level level, BlockState state, BlockPos pos, Entity entity, float fallDistance) {
-		entity.causeFallDamage(fallDistance, 0.2F, level.damageSources().fall());
+	public void fallOn(Level level, BlockState state, BlockPos pos, Entity entity, double fallDistance) {
+		entity.causeFallDamage((float) fallDistance, 0.2F, level.damageSources().fall());
 	}
 
 	@Override
